edit user availability modal in daily view
