+++ scripts/extensions/availability-manager/src/settings/availability-settings.tsx
@@ -1,13 +1,12 @@
 /* eslint-disable react/no-multi-comp */
 
 import * as React from 'react';
-import {ReactNode} from 'react';
 import * as ReactDOM from 'react-dom';
 import {addMonths, format, startOfMonth, endOfMonth} from 'date-fns';
 import {keyBy, range} from 'lodash';
 import {MonthCalendar, nameof, showModal, Spacer, SpacerBlock} from '@sourcefabric/common';
 import {Button, Card, Checkbox, getTextColor, IconButton, PopupPositioner} from 'superdesk-ui-framework/react';
-import {ISuperdeskQuery, IUser, IUserProfileSection} from 'superdesk-api';
+import {ISuperdeskQuery, IUser} from 'superdesk-api';
 import {superdesk} from '../superdesk';
 import {IAvailabilityRecord, IAvailabilityRecordTemplate, IDefaultAvailability} from '../interfaces';
 import {WorkingDayView} from './working-day-view';
@@ -26,22 +25,9 @@ const WithAvailabilityRecordsQuery = superdesk.components.getLiveQueryHOC<IAvail
 
 const verticalSpacing = '1.6rem';
 
-interface IPropsPage {
-    children: ReactNode;
-    'data-test-id'?: string;
+interface IProps {
+    user: IUser;
 }
-
-const Page: React.ComponentType<{children: React.ReactNode}> = (props: IPropsPage) => (
-    <div style={{display: 'flex', justifyContent: 'center'}} data-test-id={props['data-test-id']}>
-        <div style={{margin: '2rem'}}>
-            <Card paddingBase="3">
-                {props.children}
-            </Card>
-        </div>
-    </div>
-);
-
-type IProps = React.ComponentProps<IUserProfileSection['component']>;
 
 interface IState {
     defaultAvailability: IDefaultAvailability | null;
@@ -194,7 +180,7 @@ export class AvailabilitySettings extends React.PureComponent<IProps, IState> {
                     };
 
                     return (
-                        <Page data-test-id="availability-settings">
+                        <>
                             <div>
                                 <Checkbox
                                     checked={this.state.defaultAvailability?.enabled ?? false}
@@ -485,7 +471,7 @@ export class AvailabilitySettings extends React.PureComponent<IProps, IState> {
                                     })()}
                                 </>
                             )}
-                        </Page>
+                        </>
                     );
                 }}
             </WithAvailabilityRecordsQuery>
